Updated Scheduler
This scheduler has merged Marcel's parsing function into a lot of Stephen's framework. 
The priority queue was not actually necessary, so most of its functionality has been removed. 
The tree data structure has now been changed to an arraylist of assignment objects.
The control should work once I've finished converting all of the hard constraints and soft constraints over to not depend on trees. I've finished most of the hard constraints but there are still a few more to do. Once that's ironed out we're pretty much all good. Output should be read in a similar way to how Stephen implemented it. 
I will finish this later today.